Fix Magick.NET v14 breaking API changes (#34172)
Magick.NET v14 introduced several breaking changes that cause build
failures after the 13.5.0 → 14.10.2 bump.

### Changes

- **`MagickGeometry`**: width/height constructor params changed from
`int` to `uint` → added `(uint)` casts
- **`MagickImage.Width`/`Height`**: return type changed from `int` to
`uint` → cast back to `int` at interface boundaries
- **`RePage()`**: renamed to `ResetPage()`
- **`Compare(image, metric, diffImage, channels)`**: overload removed;
diff image is now returned directly from `Compare(image,
CompareSettings, channels, out distortion)`

```csharp
// Before
MagickImage magickDiffImage = new MagickImage();
magickDiffImage.Format = MagickFormat.Png;
double pct = baseline.Compare(actual, _errorMetric, magickDiffImage, Channels.Red);

// After
var settings = new CompareSettings(_errorMetric);
using var magickDiffImage = baseline.Compare(actual, settings, Channels.Red, out _);
magickDiffImage.Format = MagickFormat.Png;
```

+++ src/TestUtils/src/VisualTestUtils.MagickNet/MagickNetVisualComparer.cs
@@ -27,7 +27,7 @@ namespace VisualTestUtils.MagickNet
             var magickBaselineImage = new MagickImage(baselineImage.Data);
             var magickActualImage = new MagickImage(actualImage.Data);
 
-            ImageSizeDifference imageSizeDifference = ImageSizeDifference.Compare(magickBaselineImage.Width, magickBaselineImage.Height, magickActualImage.Width, magickActualImage.Height);
+            ImageSizeDifference imageSizeDifference = ImageSizeDifference.Compare((int)magickBaselineImage.Width, (int)magickBaselineImage.Height, (int)magickActualImage.Width, (int)magickActualImage.Height);
             if (imageSizeDifference != null)
                 return imageSizeDifference;
